feat(openapi): change query param type to string
Change the type of the `query` query parameter from object to string.
This is because the `python-fastapi` generator doesn't support query parameters of type object.
While OpenAPI 2 does not support query parameters of type object, OpenAPI 3 actually does support them.
However, in practice many generators are still not supporting it because they don't know how to deserialize the object from the query string.